refactor get_import_path
