+++ pyironflow/reactflow.py
@@ -2,6 +2,7 @@ from pyiron_workflow import Workflow
 from pyiron_workflow.channels import NotData
 from pyironflow.themes import get_color
 from pyironflow.create_macro import custom
+from pyironflow.utilities import get_import_path
 
 import anywidget
 import pathlib
@@ -32,16 +33,6 @@ class ReactFlowWidget(anywidget.AnyWidget):
     selected_nodes = traitlets.Unicode('[]').tag(sync=True)
     selected_edges = traitlets.Unicode('[]').tag(sync=True)
     commands = traitlets.Unicode('[]').tag(sync=True)
-
-
-def get_import_path(obj):
-    module = obj.__module__ if hasattr(obj, "__module__") else obj.__class__.__module__
-    # name = obj.__name__ if hasattr(obj, "__name__") else obj.__class__.__name__
-    name = obj.__name__ if "__name__" in dir(obj) else obj.__class__.__name__
-    path = f"{module}.{name}"
-    if path == "numpy.ndarray":
-        path = "numpy.array"
-    return path
 
 
 def dict_to_node(dict_node):
